Handle video selected
The first thing that I did is create a new state property (selectedVideo)
to know what is the current video that the user select from its search,
at the beginning this property is "null" because the user doesn't select
its video yet.
To know which video selected the user, I made the method
"handleVideoSelect()" and I pass it into the react tree until we reach
the VideoItem component. This last component will perform this method
everytime when the user clicks a video item and update the "selectedVideo" property.

Another thing that I did was render the VideoDetail component in the Video component to
display it. Also I did some styles to change the layout of the
VideoDetail component and the VideoList component.

Finally, everytime that the user make a new search, I stop to display
the VideoDetail component if the user already select a video to watch